wdpost: move fault/recover code to a separate file
